source/detectionService/requirements.txt
fastapi
uvicorn[standard]
ultralytics
lap
paho-mqtt